Mini refactor checklist->quickwit-cli
Rationale:
- checklist is only used by quickwit-cli.
- checklist pulls the colored dependency.
