+++ quickwit/quickwit-cli/src/checklist.rs
