+++ quickwit/quickwit-cli/src/split.rs
@@ -23,13 +23,13 @@ use anyhow::{bail, Context};
 use clap::{arg, ArgMatches, Command};
 use colored::Colorize;
 use itertools::Itertools;
-use quickwit_common::GREEN_COLOR;
 use quickwit_metastore::{Split, SplitState};
 use quickwit_serve::ListSplitsQueryParams;
 use tabled::{Table, Tabled};
 use time::{format_description, Date, OffsetDateTime, PrimitiveDateTime};
 use tracing::debug;
 
+use crate::checklist::GREEN_COLOR;
 use crate::{client_args, make_table, prompt_confirmation, ClientArgs};
 
 pub fn build_split_command() -> Command {
